[frontend] Lint (#5844)
